# src/aspnetcore/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "contact" : {
      "email" : "support@salestim.com",
      "name" : "SalesTim Support",
      "url" : "https://developers.salestim.com/api"
    },
    "description" : "The Microsoft Teams Governance API\n",
    "license" : {
      "name" : "(c) Copyright 2019-2020 SalesTim - all rights reserved.",
      "url" : "https://www.salestim.com/legal/eula"
    },
    "termsOfService" : "https://www.salestim.com/legal/tos",
    "title" : "SalesTim API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "SalesTim API v1.0 (Production environment, uses live data)",
    "url" : "https://api.salestim.io/v1.0"
  } ],
...
  "paths" : {
...
    "/catalog/templates" : {
      "get" : {
        "description" : "Get all templates from your corporate catalog",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/catalog/#getCatalogTemplates"
        },
        "operationId" : "GetCatalogTemplates",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArrayOfCatalogTemplates"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        }, {
          "pluginId" : [ ],
          "pluginSecret" : [ ]
        } ],
        "summary" : "Get all templates from your corporate catalog",
        "tags" : [ "Catalog" ]
      }
    },
    "/catalog/templates/installFromStore" : {
      "post" : {
        "description" : "Install a template from the public template store to your corporate catalog",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/catalog/#installTemplateFromStore"
        },
        "operationId" : "InstallTemplateFromStore",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/inline_object",
          "content" : {
            "application/json" : {
              "schema" : {
                "description" : "Object defining the ID of the template to be installed",
                "properties" : {
                  "id" : {
                    "description" : "The template ID to be installed",
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "description" : "The template to be installed",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CatalogTemplate"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Install a template from the public template store to your corporate catalog",
        "tags" : [ "Catalog" ]
      }
    },
...
    "/jobs/{id}" : {
      "get" : {
        "description" : "Get detailed information about a job (Status, logs...)",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/jobs/#getJob"
        },
        "operationId" : "GetJob",
        "parameters" : [ {
          "description" : "The job ID",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Job"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "security" : [ {
          "pluginId" : [ ],
          "pluginSecret" : [ ]
        } ],
        "summary" : "Get detailed information about a job (Status, logs...)",
        "tags" : [ "Jobs" ]
      }
    },
    "/jobs/provisioning" : {
      "post" : {
        "description" : "Create a new provisioning job by sending a ProvisioningRequest",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/jobs/#createProvisioningJob"
        },
        "operationId" : "CreateProvisioningJob",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProvisioningRequest"
              }
            }
          },
          "description" : "A ProvisioningRequest object describing the job to execute",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Job"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "security" : [ {
          "pluginId" : [ ],
          "pluginSecret" : [ ]
        } ],
        "summary" : "Create a new provisioning job by sending a ProvisioningRequest",
        "tags" : [ "Jobs" ]
      }
    }
  },
  "components" : {
    "requestBodies" : {
...
      "inline_object" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/inline_object"
            }
          }
        },
        "required" : true
      },
...
    },
...
    "schemas" : {
...
      "ArrayOfCatalogTemplates" : {
        "description" : "A list of templates from the corporate catalog",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/catalog"
        },
        "items" : {
          "$ref" : "#/components/schemas/CatalogTemplate"
        },
        "type" : "array"
      },
      "CatalogTemplate" : {
        "description" : "A template from the corporate catalog",
        "example" : {
          "templateConfiguration" : {
            "id" : "id"
          }
        },
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/catalog"
        },
        "properties" : {
          "templateConfiguration" : {
            "$ref" : "#/components/schemas/CatalogTemplate_templateConfiguration"
          }
        }
      },
...
      "Job" : {
        "description" : "A job that performs some tasks asynchronously",
        "example" : {
          "data" : "{}",
          "log" : [ "log", "log" ],
          "progress" : 8,
          "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "type" : "type",
          "status" : "active"
        },
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/jobs"
        },
        "properties" : {
          "id" : {
            "description" : "The job ID, auto-generated at creation",
            "format" : "uuid",
            "type" : "string"
          },
          "log" : {
            "description" : "The job audit log",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "type" : {
            "description" : "The job type (for instance `ProvisioningRequest`)",
            "type" : "string"
          },
          "status" : {
            "description" : "The job current status",
            "enum" : [ "active", "waiting", "completed", "failed", "deleyed", "paused" ],
            "type" : "string"
          },
          "progress" : {
            "description" : "The job current progress",
            "maximum" : 100,
            "minimum" : 0,
            "type" : "integer"
          },
          "data" : {
            "description" : "The job data (for instance a `ProvisioningRequest` object)",
            "type" : "object"
          }
        }
      },
      "ProvisioningRequest" : {
        "description" : "A request to provision a team based on a template from your corporate catalog",
        "example" : {
          "requestedOwners" : [ {
            "upn" : "upn",
            "id" : "id",
            "email" : "email"
          }, {
            "upn" : "upn",
            "id" : "id",
            "email" : "email"
          } ],
          "requestedTeamName" : "requestedTeamName",
          "requestedTeamDescription" : "requestedTeamDescription",
          "onBehalfOfRequester" : {
            "upn" : "upn",
            "id" : "id",
            "email" : "email"
          },
          "requestedTeamWelcomeMessage" : "requestedTeamWelcomeMessage",
          "templateId" : "templateId",
          "requestedMembers" : [ {
            "upn" : "upn",
            "id" : "id",
            "email" : "email"
          }, {
            "upn" : "upn",
            "id" : "id",
            "email" : "email"
          } ]
        },
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/jobs"
        },
        "properties" : {
          "templateId" : {
            "description" : "The template ID",
            "type" : "string"
          },
          "requestedTeamName" : {
            "description" : "Requested team name",
            "type" : "string"
          },
          "requestedTeamDescription" : {
            "description" : "Requested team description",
            "type" : "string"
          },
          "requestedTeamWelcomeMessage" : {
            "description" : "Requested team welcome message",
            "type" : "string"
          },
          "requestedMembers" : {
            "description" : "Requested members as an array of objects",
            "items" : {
              "$ref" : "#/components/schemas/ProvisioningRequest_requestedMembers"
            },
            "type" : "array"
          },
          "requestedOwners" : {
            "description" : "Requested owners as an array of users. User identifier could be its ID, UPN or email.",
            "items" : {
              "$ref" : "#/components/schemas/ProvisioningRequest_requestedMembers"
            },
            "type" : "array"
          },
          "onBehalfOfRequester" : {
            "$ref" : "#/components/schemas/ProvisioningRequest_onBehalfOfRequester"
          }
        },
        "required" : [ "requestedTeamName", "templateId" ]
      },
      "inline_object" : {
        "description" : "Object defining the ID of the template to be installed",
        "properties" : {
          "id" : {
            "description" : "The template ID to be installed",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CatalogTemplate_templateConfiguration" : {
        "description" : "Main template properties",
        "example" : {
          "id" : "id"
        },
        "properties" : {
          "id" : {
            "description" : "Template ID",
            "type" : "string"
          }
        }
      },
...
      "ProvisioningRequest_requestedMembers" : {
        "example" : {
          "upn" : "upn",
          "id" : "id",
          "email" : "email"
        },
        "properties" : {
          "id" : {
            "description" : "User ID",
            "type" : "string"
          },
          "upn" : {
            "description" : "User UPN",
            "type" : "string"
          },
          "email" : {
            "description" : "User email",
            "type" : "string"
          }
        }
      },
      "ProvisioningRequest_onBehalfOfRequester" : {
        "description" : "When creating a new request in `application mode`, indicates that the request should be executed on behalf of a specific user. User identifier could be its ID, UPN or email. This property is ignored in `delegated mode`.",
        "example" : {
          "upn" : "upn",
          "id" : "id",
          "email" : "email"
        },
        "properties" : {
          "id" : {
            "description" : "User ID",
            "type" : "string"
          },
          "upn" : {
            "description" : "User UPN",
            "type" : "string"
          },
          "email" : {
            "description" : "User email",
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "A valid Microsoft 365 access_token issued by Azure AD",
        "scheme" : "bearer",
        "type" : "http"
      },
      "pluginId" : {
        "description" : "A valid plugin ID",
        "in" : "header",
        "name" : "X-APP-ID",
        "type" : "apiKey"
      },
      "pluginSecret" : {
        "description" : "A valid plugin secret",
        "in" : "header",
        "name" : "X-API-KEY",
        "type" : "apiKey"
      }
    }
  }
}